+++ decidim-initiatives/spec/models/decidim/initiative_spec.rb
@@ -277,6 +277,24 @@ module Decidim
       end
     end
 
+    # TODO: OSP PR #969: Find a way to sort by support_counts
+    # describe "sorting" do
+    #  subject(:sorter) { described_class.ransack("s" => "supports_count desc") }
+
+    #  before do
+    #    create(:initiative, organization: organization, signature_type: "offline")
+    #    create(:initiative, organization: organization, signature_type: "offline", offline_votes: 4)
+    #    create(:initiative, organization: organization, signature_type: "online", initiative_votes_count: 1, initiative_supports_count: 4)
+    #    create(:initiative, organization: organization, signature_type: "online", initiative_votes_count: 2, initiative_supports_count: 1)
+    #    create(:initiative, organization: organization, signature_type: "any", initiative_votes_count: 1, initiative_supports_count: 0)
+    #    create(:initiative, organization: organization, signature_type: "any", initiative_votes_count: 3, initiative_supports_count: 2, offline_votes: 3)
+    #  end
+
+    #  it "sorts initiatives by supports count" do
+    #    expect(sorter.result.map(&:supports_count)).to eq([8, 5, 4, 3, 1, 0])
+    #  end
+    # end
+
     describe "#votes_enabled?" do
       subject { initiative.votes_enabled? }